microservice/requirements.txt
flask
flask_cors
psycopg2-binary~=2.8.6